Add default state when no stock is selected and style updates
Added a new component to display a default state when no stock is selected from the dropdown menu. This involved creating a new file, 'NoStockSelected.js', and adding logic to 'App.js' to conditionally render it. Moreover, some CSS rules were added to 'body.css' to accommodate the styling of the new state and a selected state for the stock chart options. Other minor adjustments in 'StockBasics.js' and 'StockChart.js' were made.